put GridBlock in BlockSwitcher to avoid circular dep
