+++ packages/tina-cms/src/components/index.ts
@@ -1,7 +1,6 @@
 export * from "./Banners/BannerBlocks";
 export * from "./Text/TextBlock";
 export * from "./InfoBox/InfoBoxBlock";
-export * from "./Grid/GridBlock";
 export * from "./Table/TableBlock";
 export * from "./LegacyInfoBox/LegacyInfoBoxBlock";
 export * from "./HeadingSeparator/HeadingSeparatorBlock";
